Update README.md
This is a Python script that performs web scraping on the Sobha Developers website for Bengaluru city. The script uses several Python libraries including `mmap`, `pydoc`, `bs4`, and `requests`. 

The first line of code imports the `PAGESIZE` constant from the `mmap` library, which is not used later in the code.

The second line of code imports the `pager` function from the `pydoc` library, which is also not used later in the code.

The third line of code imports the `BeautifulSoup` class from the `bs4` library, which is used for parsing HTML content.

The fourth line of code imports the `requests` library, which is used for making HTTP requests.

The next few lines of code set the URL to be scraped, send a GET request to the URL, and retrieve the HTML content of the webpage. The HTML content is then parsed using `BeautifulSoup`.

The `title` variable is set to the title of the webpage using the `soup.title` method. The title is printed to the console using the `print()` function.

The next few lines of code iterate over all the `img` tags in the HTML content and print the `src` attribute of each image to the console.

The `paras` variable is set to all the `p` tags in the HTML content using the `soup.find_all("p")` method. The `classes` variable is set to a list containing the class name `"second-tabcontent"`.

The next few lines of code iterate over all the tags with class `"second-tabcontent"` in the HTML content, extract the text content of each tag, and append it to a list called `class_content`. The `class_content` list is then converted to a string called `my_string` using the `join()` method, and printed to the console.

The `class_name` variable is then set to `"third-tabcontent"`. The script repeats the same process as above, this time extracting text content from all the tags with class `"third-tabcontent"` in the HTML content, appending it to a list called `class_content`, and converting it to a string called `my_string_1`, which is then printed to the console.

The script then attempts to extract phone numbers from the HTML content using a regular expression. The regular expression `\b\d{12}\b` matches 12 consecutive digits that are surrounded by word boundaries. The script uses the `re.findall()` method to extract all matches of this regular expression in the HTML content. The phone numbers are then printed to the console using a `for` loop.

The next few lines of code are commented out, but they appear to extract all the links on the webpage and print the relevant links to the console. However, these lines of code will not work correctly since the `links` variable has not been defined.

In summary, this script performs basic web scraping on the Sobha Developers website for Bengaluru city. It extracts the title of the webpage, the `src` attribute of all images, and the text content of all tags with class `"second-tabcontent"` and `"third-tabcontent"`. It also attempts to extract phone numbers from the HTML content using a regular expression. However, the commented out code to extract all links on the webpage will not work correctly.